feat: add all layouts

+++ lib/random.hpp
@@ -3,7 +3,7 @@
 
 class Random {
   public:
-    static ulong Rng(const int &floor, const int &ceiling) {
+    static ulong Rng(const ulong &floor, const ulong &ceiling) {
         std::random_device dev;
         std::mt19937 rng(dev());
         std::uniform_int_distribution<std::mt19937::result_type> dist(floor,
